update readme, don't allow db:seed in prod

+++ db/seeds.rb
@@ -1,3 +1,5 @@
+return if Rails.env.production?
+
 User.destroy_all
 
 user = User.create!(
